:sparkles: Added ViewFileTrait containing methods shared between the ViewFile class and the Plates view helpers extension
