+++ src/Entities/ViewFile.php
@@ -10,6 +10,9 @@ namespace Tapestry\Entities;
  */
 class ViewFile
 {
+
+    use ViewFileTrait;
+
     /**
      * @var Project
      */
@@ -43,67 +46,9 @@ class ViewFile
     public function getFile()
     {
         if (is_null($this->file)) {
-            $this->file = $this->project->get('compiled.'.$this->fileUid);
+            $this->file = $this->project->get('compiled.' . $this->fileUid);
         }
 
         return $this->file;
-    }
-
-    public function getData($key, $default = null)
-    {
-        return $this->getFile()->getData($key, $default);
-    }
-
-    public function getPermalink()
-    {
-        return $this->getFile()->getCompiledPermalink();
-    }
-
-    public function getUrl()
-    {
-        return url($this->getPermalink());
-    }
-
-    /**
-     * @return \DateTime
-     */
-    public function getDate()
-    {
-        return $this->getData('date');
-    }
-
-    public function getContent()
-    {
-        if ($content = $this->getData('content')) {
-            return $content;
-        }
-
-        return $this->getFile()->getContent();
-    }
-
-    public function isPaginated()
-    {
-        if (! $pagination = $this->getData('pagination')) {
-            return false;
-        }
-
-        if (! $pagination instanceof Pagination) {
-            return false;
-        }
-
-        return true;
-    }
-
-    public function hasPreviousNext()
-    {
-        if (! $previousNext = $this->getData('previous_next')) {
-            return false;
-        }
-
-        if (! $previousNext instanceof \stdClass) {
-            return false;
-        }
-
-        return true;
     }
 }
